requirements.txt
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==2.2.9
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.3
Markdown==3.3.6
Pillow==9.1.1
psycopg2==2.8.6
pycodestyle==2.8.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
zipp==3.8.0
django_heroku==0.3.1
gunicorn==19.9.0
cloudinary==1.29.0